requirements.txt
torch
torchvision
numpy
scikit-learn
tqdm
tensorboard
matplotlib
Pillow
